Update Examples to show "Stacked" series charts
