+++ examples/README.md
@@ -6,15 +6,18 @@ Examples are our primary method for demonstrating the starting point our use and
 
 * [bar_chart-1](./bar_chart-1) - Bar chart with included mark points and mark lines.
 * [bar_chart-2](./bar_chart-2) - The above example bar chart re-demonstrated using the Painter API.
+* [bar_chart-3](./bar_chart-3) - A bar chart with "Stacked" series enabled, collapsing the bars into a single layered bar.
 * [chinese](./chinese) - Line chart with chinese characters that uses a custom font (must be downloaded by user, see comment in code).
 * [funnel_chart-1](./funnel_chart-1) - Basic funnel chart.
 * [funnel_chart-2](./funnel_chart-2) - The above example funnel chart re-demonstrated using the Painter API.
 * [horizontal_bar_chart-1](./horizontal_bar_chart-1) - Basic horizontal bar chart.
 * [horizontal_bar_chart-2](./horizontal_bar_chart-2) - The above example bar chart re-demonstrated using the Painter API.
+* [horizontal_bar_chart-3](./horizontal_bar_chart-3) - A horizontal bar chart with "Stacked" series, collapsing the bars into a single layered bar.
 * [line_chart-1](./line_chart-1) - Basic line chart with some simple styling changes and a demonstration of `null` values.
 * [line_chart-2](./line_chart-2) - The above example line chart re-demonstrated using the Painter API.
 * [line_chart-3](./line_chart-3) - Line chart with dense data and more custom styling configured.
-* [line_chart-4](./line_chart-4) - Line chart with dense data and most default rendering disabled, instead rendering labels manually on the Painter.
+* [line_chart-4](./line_chart-4) - Line chart with "Stacked" series enabled, making each series a layer on the chart and the top line showing the sum.
+* [line_chart-5](./line_chart-5) - Line chart with dense data and most default rendering disabled, instead rendering labels manually on the Painter.
 * [line_chart-area](./line_chart-area) - Example line chart with the area below the line shaded.
 * [multiple_charts-1](./multiple_charts-1) - Example of manually building a painter so that you can render 4 charts on the same image.
 * [multiple_charts-2](./multiple_charts-2) - Combining two charts together by writting one chart over the other.
